Move AuditService.Write() method to its own service
