+++ src/Umbraco.Core/Services/OperationStatus/AuditEntryOperationStatus.cs
@@ -1,0 +1,10 @@
+namespace Umbraco.Cms.Core.Services.OperationStatus;
+
+/// <summary>
+/// Represents the status of an audit entry operation.
+/// </summary>
+public enum AuditEntryOperationStatus
+{
+    Success,
+    RepositoryNotAvailable,
+}
